+++ packages/snaps-sdk/src/internals/helpers.ts
@@ -1,3 +1,5 @@
+import type { JsonRpcParams, JsonRpcRequest } from '@metamask/utils';
+
 /**
  * Get the enum values as union type. This allows using both the enum string
  * values and the enum itself as values.
@@ -5,17 +7,35 @@
  * Note: This only works for string enums.
  *
  * @example
- * ```typescript
  * enum Foo {
  *   Bar = 'bar',
  *   Baz = 'baz',
  * }
  *
- * type FooValue = EnumToUnion<Foo>;
  * // FooValue is 'bar' | 'baz'
+ * type FooValue = EnumToUnion<Foo>;
  *
  * const foo: FooValue = Foo.Bar; // Works
  * const foo: FooValue = 'bar'; // Also works
- * ```
  */
 export type EnumToUnion<Type extends string> = `${Type}`;
+
+/**
+ * Get a JSON-RPC method with the given name and parameters. If params extends
+ * `never`, then the `params` property is omitted.
+ *
+ * @example
+ * // MyMethod is { method: 'my_method', params: { foo: string } }
+ * type MyMethod = Method<'my_method', { foo: string }>;
+ */
+export type Method<
+  MethodName extends string,
+  Params extends JsonRpcParams,
+> = Partial<JsonRpcRequest> & Params extends never
+  ? {
+      method: MethodName;
+    }
+  : {
+      method: MethodName;
+      params: Params;
+    };
