+++ packages/snaps-sdk/src/types/methods/dialog.ts
@@ -25,3 +25,24 @@ export type PromptDialog = {
 };
 
 export type Dialog = AlertDialog | ConfirmationDialog | PromptDialog;
+
+/**
+ * The request parameters for the `snap_dialog` method.
+ *
+ * @property type - The type of dialog to display.
+ * @property content - The content to display in the dialog.
+ * @property placeholder - The placeholder text to display in the dialog. Only
+ * applicable for the `prompt` dialog.
+ */
+export type DialogParams = Dialog;
+
+/**
+ * The result returned by the `snap_dialog` method.
+ *
+ * - If the dialog is an `alert`, the result is `null`.
+ * - If the dialog is a `confirmation`, the result is a boolean indicating
+ * whether the user confirmed the dialog.
+ * - If the dialog is a `prompt`, the result is the value entered by
+ * the user.
+ */
+export type DialogResult = null | boolean | string;
