+++ packages/snaps-sdk/src/types/methods/get-file.ts
@@ -1,5 +1,23 @@
+import type { EnumToUnion } from '../../internals';
+
 export enum AuxiliaryFileEncoding {
   Base64 = 'base64',
   Hex = 'hex',
   Utf8 = 'utf8',
 }
+
+/**
+ * The request parameters for the `snap_getFile` method.
+ *
+ * @property path - The path to the file to retrieve.
+ * @property encoding - The encoding to use when retrieving the file.
+ */
+export type GetFileParams = {
+  path: string;
+  encoding?: EnumToUnion<AuxiliaryFileEncoding>;
+};
+
+/**
+ * The result returned by the `snap_getFile` method.
+ */
+export type GetFileResult = string;
