+++ packages/snaps-sdk/src/types/methods/get-locale.ts
@@ -1,0 +1,13 @@
+/**
+ * The request parameters for the `snap_getLocale` method.
+ *
+ * This method does not accept any parameters.
+ */
+export type GetLocaleParams = never;
+
+/**
+ * The result returned by the `snap_getLocale` method.
+ *
+ * It is the locale of the user's MetaMask extension.
+ */
+export type GetLocaleResult = string;
