+++ packages/snaps-sdk/src/types/methods/notify.ts
@@ -9,3 +9,18 @@ export type Notify = {
   type: EnumToUnion<NotificationType>;
   message: string;
 };
+
+/**
+ * The request parameters for the `snap_notify` method.
+ *
+ * @property type - The type of notification to display.
+ * @property message - The message to display in the notification.
+ */
+export type NotifyParams = Notify;
+
+/**
+ * The result returned by the `snap_notify` method.
+ *
+ * This method does not return anything.
+ */
+export type NotifyResult = null;
